+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/indentation/commentsindentation/InputCommentsIndentationSurroundingCode3One.java
@@ -1,6 +1,6 @@
 /*
 CommentsIndentation
-tokens = SINGLE_LINE_COMMENT
+tokens = BLOCK_COMMENT_BEGIN
 
 
 */
@@ -11,13 +11,12 @@ package com.puppycrawl.tools.checkstyle.checks.indentation.commentsindentation;
 import java.util.Arrays;
 
 // some
-public class InputCommentsIndentationSurroundingCode2
+public class InputCommentsIndentationSurroundingCode3One
 {
     private void foo1() {
         if (true) {
             // here initialize some variables
             int k = 0; // trailing comment
-              // violation '.* incorrect .* level 14, expected is 12, .* same .* as line 21.'
             int b = 10;
             // sss
         }
@@ -27,12 +26,13 @@ public class InputCommentsIndentationSurroundingCode2
         if (true) {
             /* some */
             int k = 0;
-                /* violation */
+            // violation below '.* incorrect .* level 16, expected is 12, .* same .* as line 31.'
+                /* some comment */
             int b = 10;
-                /* violation
+                /* // violation '.* incorrect .* level 16, expected is 12, .* same .* as line 34.'
                  * */
             double d; /* trailing comment */
-                /* violation
+                /* // violation '.* incorrect .* level 16, expected is 12, .* same .* as line 38.'
              *
                 */
             boolean bb;
@@ -54,8 +54,7 @@ public class InputCommentsIndentationSurroundingCode2
         int a = 5, b = 3, v = 6;
         if (a == b
             && v == b || ( a ==1
-                           /// violation '.* incorrect .* level 27, expected is 36, .* as line 60.'
-                       /* violation
+                       /* // violation '.* incorrect .* level 23, expected is 36, .* as line 59.'
                         * one fine day ... */
                                     && b == 1)   ) {
         }
@@ -94,7 +93,6 @@ public class InputCommentsIndentationSurroundingCode2
               // ...
               // block
               // ...
-              // violation '.* incorrect .* level 14, expected is 8, .* same .* as line 98.'
         String someStr = new String();
     }
 
@@ -102,62 +100,8 @@ public class InputCommentsIndentationSurroundingCode2
              // comment
              // ...
              // block
-             // violation '.* incorrect .* level 13, expected is 8, .* same .* as line 107.'
         // comment
         String someStr = new String();
-    }
-
-    public void foo8() {
-        String s = new String(); // comment
-                                 // ...
-                                 // block
-                                 // ...
-                                 // violation '.*incorrect.*level 33, expected is 8,.*as line 116.'
-        String someStr = new String();
-    }
-
-
-    public String foo9(String s1, String s2, String s3) {
-        return "";
-    }
-
-    public void foo10()
-        throws Exception {
-
-        final String pattern = "^foo$";
-
-        final String[] expected = {
-            "7:13: " + foo9("", "", ""),
-            // comment
-        };
-    }
-
-    public void foo11() {
-
-            /* empty */
-        hashCode();
-    }
-
-    public void foo12() {
-    /* empty */
-        hashCode();
-    }
-
-    public void foo13() {
-        hashCode();
-    /* empty */
-    }
-
-    public void foo14() {
-        hashCode();
-        /*
-
-        Test
-        */
-        // Test
-    }
-
-    public InputCommentsIndentationSurroundingCode2() {
     }
 
     // Test
